rationalising code structure
